pyproject.toml
[project]
name = "591-crawler"
version = "0.1.0"
description = "591 rental listing crawler with notification system"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Web Framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    # Browser Automation
    "playwright>=1.40.0",
    # HTTP Client
    "httpx>=0.27.0",
    # PostgreSQL
    "asyncpg>=0.29.0",
    "sqlalchemy[asyncio]>=2.0.0",
    # Redis
    "redis>=5.0.0",
    # Scheduler
    "apscheduler>=3.10.0",
    # Telegram Bot
    "python-telegram-bot>=21.0",
    # Configuration
    "pydantic-settings>=2.0.0",
    # Logging
    "loguru>=0.7.0",
    # Environment
    "python-dotenv>=1.0.0",
    "email-validator>=2.3.0",
    "pyjwt>=2.10.1",
]

[tool.uv]
dev-dependencies = [
    # Testing
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    # Linting & Formatting
    "ruff>=0.2.0",
    # Type checking
    "mypy>=1.8.0",
]

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
